replace venus api by private extensions
Having it as extensions makes it possible to extend existing Vulkan
enums and pnext structs.

VK_EXT_command_serialization defines the wire format and how Vulkan
commands are serialized.  It is an extension because it might extend
external memory/fence/semaphore enums and pnext structs.

VK_MESA_venus_protocol defines new Vulkan commands, enums, and structs
for venus's own use.  It is an extension because it might extend
VkQueueFamilyProperties2, etc.