rework VkVariable::is_string
Rename it to has_c_string because it returns true for an array of
C-strings as well.

We can work out the answer from xml attributes or from the C type.
Because this is a VkVariable method than a VkType one, and is more
high-level, use the xml attributes.

No change to the generated code.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>