handle arrays with user defined strides
vkCmdDrawMultiEXT/vkCmdDrawMultiIndexedEXT pass in arrays with
user defined strides between each element.

This wasn't handled by vn_protocol.py so it was reading from the wrong
offsets, expecting the array to be packed with standard strides of
sizeof(element).

This change adds a special case for arrays with strides when variables
from vk.xml have the 'stride' attribute.

The values are encoded packed and the stride is replaced with the
standard value so the decoder handles this as a normal array and does
not require modification.

Note: for the special case of stride=0, multiple copies of the element
are encoded.

Signed-off-by: Juston Li <justonli@chromium.org>