Handle conditionally ignored dynamic arrays
Previously, the protocol simply checks if a pointer is non-NULL before
encoding dynamic arrays. In certain cases when the pointer should be
ignored, venus will incorrectly read and encode from the non-NULL
pointer.

Conditions for which/when an arg should be ignored isn't readily found
in the xml so hardcode a list of possibly ignored variables along with
the condition that they should be ignored and replace the check before
encoding dynamic arrays.

Currently, that includes the conditionally ignored arg
pQueueFamilyIndices found in the 3 structs VkImageCreateInfo,
VkBufferCreateInfo, VkPhysicalDeviceImageDrmFormatModifierInfoEXT along
with VkFramebufferCreateInfo->pAttachments.

Signed-off-by: Juston Li <justonli@google.com>
Part-of: <https://gitlab.freedesktop.org/virgl/venus-protocol/-/merge_requests/91>