New protocol support for VK_EXT_host_image_copy
To best leverage in-place blob encoding which avoids the temp
allocation and the memcpy to the reply shmem, the custom protocol is
designed to flatten the img2mem copy. That means we have a bit more cs
overhead when there're multiple regions to copy, but the gain is much
bigger by not allocating image-sized data along with memcpy. Meanwhile,
we can still keep the nested mem2img copy as is since the in-place blob
decoding doesn't need excessive offset calculation of the reply shmem.

Signed-off-by: Yiwei Zhang <zzyiwei@chromium.org>
Part-of: <https://gitlab.freedesktop.org/virgl/venus-protocol/-/merge_requests/95>