fix in-place blob decoding
This doesn't affect existing extensions supported, but was found during
adding custom support for host image copy. We still have to do temp
allocs for blobs if the out blob is hidden inside a struct. The easy way
to handle this is to only use in-place blob decoding when the blob is a
const pointer.

Signed-off-by: Yiwei Zhang <zzyiwei@chromium.org>
Part-of: <https://gitlab.freedesktop.org/virgl/venus-protocol/-/merge_requests/95>